refactor: 리팩토링 & 스타일 변경
