+++ src/main/java/com/programmers/util/Menu.java
@@ -1,6 +1,7 @@
 package com.programmers.util;
 
 import com.programmers.exception.MenuFormatException;
+import com.programmers.model.Request;
 
 import java.util.Arrays;
